WIP: Collapsed ODM definitions into a single file getting rid off database module
