--- metadataStore/database/__init__.py
@@ -1,7 +1,0 @@
-from __future__ import (absolute_import, division, print_function)
-
-__author__ = 'arkilic'
-from .beamline_config import BeamlineConfig
-from .event_descriptor import EventDescriptor
-from .event import Event
-from .begin_run_event import Header
